# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Diagnostic": {
        "description": "A warning, error or translation notice about a query, with a half-open code-point span into `q`.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/DiagnosticCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "span": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span"
          }
        },
        "required": [
          "code",
          "message",
          "span"
        ],
        "title": "Diagnostic",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "Excluded": {
        "description": "03's exclusion accounting, verbatim: buckets by count (largest first, ties by name), `unknown` last.",
        "properties": {
          "status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "track": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Track",
            "type": "object"
          }
        },
        "required": [
          "total",
          "track",
          "status"
        ],
        "title": "Excluded",
        "type": "object"
      },
      "Facets": {
        "description": "Disjunctive (decision-001): each field counted without its own top-level conjuncts. Values that\noccur only; years as strings.",
        "properties": {
          "status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status",
            "type": "object"
          },
          "track": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Track",
            "type": "object"
          },
          "venue": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Venue",
            "type": "object"
          },
          "year": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Year",
            "type": "object"
          }
        },
        "required": [
          "venue",
          "year",
          "track",
          "status"
        ],
        "title": "Facets",
        "type": "object"
      },
...
      "Hit": {
        "properties": {
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "highlights": {
            "$ref": "#/components/schemas/Highlights"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "presentation": {
            "anyOf": [
              {
                "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
                "enum": [
                  "oral",
                  "spotlight",
                  "poster"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presentation"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "status": {
            "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
            "enum": [
              "accepted",
              "rejected",
              "withdrawn",
              "desk_rejected",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "track": {
            "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
            "enum": [
              "main",
              "datasets_benchmarks",
              "position",
              "workshop",
              "competition",
              "tiny_papers",
              "blogpost",
              "other",
              "unknown"
            ],
            "title": "Track",
            "type": "string"
          },
          "urls": {
            "$ref": "#/components/schemas/Urls"
          },
          "venue": {
            "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
            "enum": [
              "NeurIPS",
              "ICLR",
              "ICML"
            ],
            "title": "Venue",
            "type": "string"
          },
          "year": {
            "minimum": 1000,
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "abstract",
          "authors",
          "venue",
          "year",
          "track",
          "status",
          "presentation",
          "score",
          "highlights",
          "urls"
        ],
        "title": "Hit",
        "type": "object"
      },
...
      "QueryInfo": {
        "properties": {
          "canonical": {
            "title": "Canonical",
            "type": "string"
          },
          "canonical_hash": {
            "title": "Canonical Hash",
            "type": "string"
          },
          "expansions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Expansions",
            "type": "object"
          },
          "identification_query": {
            "title": "Identification Query",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Translations",
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "input",
          "canonical",
          "canonical_hash",
          "identification_query",
          "warnings",
          "translations",
          "expansions"
        ],
        "title": "QueryInfo",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "excluded": {
            "$ref": "#/components/schemas/Excluded"
          },
          "facets": {
            "$ref": "#/components/schemas/Facets"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/Hit"
            },
            "title": "Hits",
            "type": "array"
          },
          "index_version": {
            "title": "Index Version",
            "type": "string"
          },
          "query": {
            "$ref": "#/components/schemas/QueryInfo"
          },
          "query_version": {
            "title": "Query Version",
            "type": "string"
          },
          "tokenizer_version": {
            "title": "Tokenizer Version",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "index_version",
          "tokenizer_version",
          "query_version",
          "query",
          "total",
          "excluded",
          "facets",
          "hits"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "openproceedings",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/export": {
      "get": {
        "description": "Every record the query matches, in `format`: of `q` (with `mode`, default `native`) on the pinned\n`index_version` (else the served index), or the stored ids of search record `record_id` (alone) from the\nindex it names. Never paginated or truncated; `X-Total` equals `/search`'s `total` for the same query and\nindex (for a record, its stored `total`).",
        "operationId": "export",
        "parameters": [
          {
            "description": "The file format.",
            "in": "query",
            "name": "format",
            "required": true,
            "schema": {
              "description": "The file format.",
              "enum": [
                "ris",
                "csv",
                "bibtex",
                "jsonl"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising. Required unless `record_id` is given.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising. Required unless `record_id` is given.",
              "title": "Q"
            }
          },
          {
            "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated). With `record_id`, only `native` (a record's canonical string is native syntax); `scholar` there is 422 `API_BAD_PARAM`.",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "native",
              "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated). With `record_id`, only `native` (a record's canonical string is native syntax); `scholar` there is 422 `API_BAD_PARAM`.",
              "enum": [
                "native",
                "scholar"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "Export from this index (409 `API_INDEX_VERSION_UNAVAILABLE` if this instance can't serve it); the served one if absent. Only with `q`.",
            "in": "query",
            "name": "index_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[0-9a-f][0-9a-f-]{0,63})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Export from this index (409 `API_INDEX_VERSION_UNAVAILABLE` if this instance can't serve it); the served one if absent. Only with `q`.",
              "title": "Index Version"
            }
          },
          {
            "description": "Export exactly this search record's stored ids, from the index it names. Not with `q` or `index_version`; `mode`, if sent, must be `native`.",
            "in": "query",
            "name": "record_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(?:[A-Za-z0-9_-]{12})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Export exactly this search record's stored ids, from the index it names. Not with `q` or `index_version`; `mode`, if sent, must be `native`.",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/x-bibtex": {},
              "application/x-ndjson": {},
              "application/x-research-info-systems": {},
              "text/csv": {}
            },
            "description": "The entire matched set, ordered by id.",
            "headers": {
              "Content-Disposition": {
                "description": "attachment; filename=\"openproceedings-<index_version>-<first 12 of canonical_hash>.<ext>\"",
                "schema": {
                  "type": "string"
                }
              },
              "X-Index-Version": {
                "description": "The index the records were read from",
                "schema": {
                  "type": "string"
                }
              },
              "X-Query-Version": {
                "description": "This code's query_version (on a `record_id` export too, whatever the record's own)",
                "schema": {
                  "type": "string"
                }
              },
              "X-Tokenizer-Version": {
                "description": "This code's tokenizer_version",
                "schema": {
                  "type": "string"
                }
              },
              "X-Total": {
                "description": "How many records the body holds: `/search`'s `total` for the same query and index (for a search record, the number of its stored ids, which its `total` must equal: 409 `API_RECORD_MISMATCH` otherwise)",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_METHOD_NOT_ALLOWED: the endpoint exists, but not for this method",
            "headers": {
              "Allow": {
                "description": "The methods this endpoint takes, comma-separated (RFC 9110 §10.2.1)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_RATE_LIMITED: this client's token bucket or its network's (IPv4 /24, IPv6 /48) can't pay for the request, a query's position-verified clauses included; or, on POST /records, the save ceiling of this client's network or of the whole instance is reached",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the request would be allowed",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_BUSY (with Retry-After): the query needs a slow position check and every verification slot is taken; or API_INDEX_NOT_LOADED (no index loaded yet), or on POST /records API_RECORDS_STORE_FULL (neither sends Retry-After)",
            "headers": {
              "Retry-After": {
                "description": "Sent with API_BUSY: whole seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error (spec 04 §Error handling)"
          }
        },
        "summary": "Export"
      }
    },
...
    "/api/v1/search": {
      "get": {
        "description": "One page of the ranked matched set, with its total, exclusion accounting, disjunctive facets and\nhighlights. `limit` over 200 is a 422 (never clamped); a query that doesn't parse is a 422 with its\ndiagnostics.",
        "operationId": "search",
        "parameters": [
          {
            "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated).",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "native",
              "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated).",
              "enum": [
                "native",
                "scholar"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "The order of `hits`. Never changes `total` or membership (guarantee 5).",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "relevance",
              "description": "The order of `hits`. Never changes `total` or membership (guarantee 5).",
              "enum": [
                "relevance",
                "year_desc",
                "year_asc",
                "title"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "Hits to skip. Past the end is an empty page, not an error.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Hits to skip. Past the end is an empty page, not an error.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Hits per page, 0 to 200. Over 200 is 422 `API_BAD_PARAM`, never clamped.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Hits per page, 0 to 200. Over 200 is 422 `API_BAD_PARAM`, never clamped.",
              "maximum": 200,
              "minimum": 0,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_METHOD_NOT_ALLOWED: the endpoint exists, but not for this method",
            "headers": {
              "Allow": {
                "description": "The methods this endpoint takes, comma-separated (RFC 9110 §10.2.1)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_RATE_LIMITED: this client's token bucket or its network's (IPv4 /24, IPv6 /48) can't pay for the request, a query's position-verified clauses included; or, on POST /records, the save ceiling of this client's network or of the whole instance is reached",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the request would be allowed",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_BUSY (with Retry-After): the query needs a slow position check and every verification slot is taken; or API_INDEX_NOT_LOADED (no index loaded yet), or on POST /records API_RECORDS_STORE_FULL (neither sends Retry-After)",
            "headers": {
              "Retry-After": {
                "description": "Sent with API_BUSY: whole seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error (spec 04 §Error handling)"
          }
        },
        "summary": "Search"
      }
    }
  }
}